Time: 3 ms (77.29%), Space: 6 MB (25.79%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -4,11 +4,11 @@ public:
         int s=0;
         int e=x;
         
-       long long int ans=-1;
+       long int ans=-1;
         
         while(s<=e)
         {
-           long long int mid=s+(e-s)/2;
+           long int mid=s+(e-s)/2;
            long long int square=mid*mid;
             if(square == x)
             {
